requirements.txt
fastapi
uvicorn
sqlalchemy
deepface
tf-keras
opencv-python
python-multipart
python-jose
passlib
bcrypt
aiofiles
supabase
psycopg2-binary
python-dotenv
